pyproject.toml
[tool.poetry]
name = "guarana"
version = "0.2.0"
description = ""
authors = ["Rodrigo Godinho <rodrigolegod@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
segment-analytics-python = "^2.2.0"
loguru = "^0.6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
